+++ internal/server/pvp.go
@@ -62,7 +62,7 @@ func (h *hub) attackPlayer(players map[int32]*tracked, attacker, target int32) b
 	h.toNearbyEv(players, t.dim, t.x, t.z, swingArm(t.p.eid))
 	switch {
 	case sw.crit:
-		h.spawnParticles(players, particleCrit, v.x, v.y+1, v.z, 0.4, 0.2, 8)
+		h.spawnParticles(players, v.dim, particleCrit, v.x, v.y+1, v.z, 0.4, 0.2, 8)
 		h.playSound(players, "minecraft:entity.player.attack.crit", sndPlayer, v.x, v.y, v.z, 1, 1)
 	case sw.charge >= 0.9:
 		h.playSound(players, "minecraft:entity.player.attack.strong", sndPlayer, v.x, v.y, v.z, 1, 1)